+++ server/router.go
@@ -26,14 +26,14 @@ func InitRoute() error {
 	routes.InitUserRoute(ginServer)
 	routes.InitProblemRoute(ginServer)
 	routes.InitTagRoute(ginServer)
+	routes.InitSolutionRoute(ginServer)
+	routes.InitTestcaseRoute(ginServer)
 	routes.InitJudgeRoute(ginServer)
 	routes.InitRecordRoute(ginServer)
-	routes.InitSolutionRoute(ginServer)
 	routes.InitBlogRoute(ginServer)
 	routes.InitCommentRoute(ginServer)
 	routes.InitAiRouter(ginServer)
 	routes.InitStatisticsRoute(ginServer)
-	routes.InitTestcaseRoute(ginServer)
 	routes.InitSystemRoute(ginServer)
 	routes.InitMiscRoute(ginServer)
